+++ examples/resources/mongodb_collection/resource.tf
@@ -1,0 +1,4 @@
+resource "mongodb_collection" "example" {
+  database = "test"
+  name     = "example"
+}