+++ internal/provider/database_resource_test.go
@@ -1,0 +1,25 @@
+package provider
+
+import (
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+)
+
+func TestAccDatabaseResource(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			{
+				Config: providerConfig + `
+resource "mongodb_database" "test_db_new" {
+	name = "test_db_new"
+}
+`,
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr("mongodb_database.test_db_new", "name", "test_db_new"),
+				),
+			},
+		},
+	})
+}