+++ internal/provider/utils.go
@@ -1,10 +1,14 @@
 package provider
 
 import (
+	"crypto/tls"
+	"crypto/x509"
 	"errors"
+	"net/url"
 	"strings"
 
 	"go.mongodb.org/mongo-driver/mongo/options"
+	"golang.org/x/net/proxy"
 )
 
 // Convert an index type declared in terraform as string into a type and value expected by Mongo's client.
@@ -89,4 +93,48 @@ func (co *collation) toMongoCollation() *options.Collation {
 		res.Backwards = *co.Backwards
 	}
 	return &res
+}
+
+func addArgs(arguments string, newArg string) string {
+	if arguments != "" {
+		return arguments + "&" + newArg
+	} else {
+		return "/?" + newArg
+	}
+
+}
+
+func getTLSConfigWithAllServerCertificates(ca []byte, verify bool) (*tls.Config, error) {
+	/* As of version 1.2.1, the MongoDB Go Driver will only use the first CA server certificate found in sslcertificateauthorityfile.
+	   The code below addresses this limitation by manually appending all server certificates found in sslcertificateauthorityfile
+	   to a custom TLS configuration used during client creation. */
+
+	tlsConfig := new(tls.Config)
+
+	tlsConfig.InsecureSkipVerify = verify
+	tlsConfig.RootCAs = x509.NewCertPool()
+	ok := tlsConfig.RootCAs.AppendCertsFromPEM(ca)
+
+	if !ok {
+		return tlsConfig, errors.New("Failed parsing pem file")
+	}
+
+	return tlsConfig, nil
+}
+
+func proxyDialer(proxyUrlFromProvider string) (options.ContextDialer, error) {
+	if proxyUrlFromProvider != "" {
+		proxyURL, err := url.Parse(proxyUrlFromProvider)
+		if err != nil {
+			return nil, err
+		}
+		proxyDialer, err := proxy.FromURL(proxyURL, proxy.Direct)
+		if err != nil {
+			return nil, err
+		}
+
+		return proxyDialer.(options.ContextDialer), nil
+	}
+
+	return nil, nil
 }