+++ CHANGELOG.md
@@ -3,3 +3,10 @@
 FEATURES:
 
 - Add index resource
+
+# 0.2.0
+
+FEATURES
+
+- Add database resource
+- Add collection resource